content/html/content/test/test_bug546995-4.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=546995
-->
<head>
  <title>Test for Bug 546995</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=546995">Mozilla Bug 546995</a>
<p id="display"></p>
<iframe id="iframe" onload='onFrameLoad();'></iframe>
<div id='content'>
</div>
<pre id="test">
<script type="application/javascript;version=1.8">

/** Test for Bug 546995 **/

SimpleTest.waitForExplicitFinish();

function onFrameLoad()
{
  document.getElementById('iframe').focus();

  var input = document.createElement('input');
  input.id = 'i';
  input.autofocus = true;
  document.getElementById('content').appendChild(input);

  // When |runTests| will be executed, the focus event from autofocus should
  // have been processed.
  SimpleTest.executeSoon(runTests);
}

function runTests()
{
  isnot(document.activeElement, document.getElementById('i'),
        "The focus should not be able to leave a sub-document");
  is(document.activeElement, document.getElementById('iframe'),
     "The iframe should have the focus");

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
